requirements.txt
fastapi
uvicorn
python-telegram-bot
langchain
langchain-openai
langchain-community
mcp
sqlalchemy
aiosqlite
pydantic-settings
httpx
python-dotenv